backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
Faker==24.11.0
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
PyJWT==2.8.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
